use AccountManagement assembly and change publishprofiles
use System.DirectoryServices.AccountManagement to change the password. Use the UserPrincipal.ChangePassword() method. Remove old code and assembly. Add 2 appSettings, DomainName and SearchBase, both of which are optional and not dependent on each other. Remove the LDAPConnectionString Connection string. Use
IPGlobalProperties.GetIPGlobalProperties().DomainName to get the default domain name.

Default the searchbase to be the same as the domain, in LDAP form, i.e.
dc=example,dc=com.

Replace the to file publish profile with the passwd_inetpub profile, which deploys it to c:\inetpub\passwd\, which is a good place to put the application if you intent to serve it as /passwd in IIS.

Remove the placeholder label "servername".